refactor data storage for hypercine to always be uint16_t type
This is regardless of the bit depth of the images read, be they 8-bit
or 16-bit. The reason is to be able to expose a pointer to a constant
type regardless of the bit depth so memory can be shared with other
applications via a simple pointer where the type is known.

The only downside to this is that 8bit images double in size, but
the simplicity of always being able to access the intensity values
via a pointer to uint16_t makes up for it.