refactor data(frame,window_id) and data(frame,x_begin,...)
now if the buffer doesn't have the data for the requested frame or
window id or window dimensions it automatically reads that data into
the buffer so long as the frame and window ids, etc. are valid.

This still enables the user to read the buffer manually (for example
to read in large chunks of the cine file) but doesn't require it.

Also added better test coverage for all these new automated reading
modes.