added convenience get_frame() to read a full frame and return a vector
this method doesn't interfere with the memory buffer of the HyperCine
class. It is intended for cases where the user wants to read a full
frame of a cine file, but may have already allocated and read a large
chunk of the memory buffer. This method returns a copied vector of
the intensity values