enable user to set storage type (no longer fixed as uint16_t) from cmake
this is to enable passing a pointer to another storage type for the
application that is using hypercine. For example the parent application
may be using float type to store the cine intensity values, but if
the storage type is uint16_t, each value needs to be copied and
casted.