moved HyperFrame methods implementations to .cpp file
for some reason, if the methods are in the header, on linux this leads
to seg faults when accessing the HyperFrame via a HyperCine. It was
giving really strange behavior where two HyperFrames were being
constructed in memory (with different addresses). This behavior was
observed when linking hypercine lib with another library that made
use of a HyperCine object.